Using Shared Context in out tests!

+++ DB.Test/SQLServerUT.cs
@@ -3,17 +3,33 @@ using System.Data.SqlClient;
 
 namespace DB.Test
 {
-    public class SQLServerUT
+    public class SQLServerUT : IDisposable
     {
+
+        SqlConnection conn;
+
+        public SQLServerUT()
+        {
+            conn = new SqlConnection("Data Source=192.168.100.30,1433;Initial Catalog=SQLServerUT;Integrated Security=SSPI;");
+            conn.Open();
+        }
+
+        public void Dispose()
+        {
+            conn.Dispose();
+        }
+
+
         [Theory]
         [InlineData(2, 3, 5)]
         [InlineData(-5, 5, 0)]
         [InlineData(-2, -5, -7)]
         public void usf_SimpleAdd_IntegerValues_Calculated(int a, int b, int expected)
         {
-            using (SqlConnection conn = new SqlConnection("Data Source=192.168.100.30,1433;Initial Catalog=SQLServerUT;Integrated Security=SSPI;"))
+
+            if (conn != null && conn.State == ConnectionState.Open)
             {
-                string sql = "SELECT dbo.usf_SimplyAdd(@a, @b);"; 
+                string sql = "SELECT dbo.usf_SimplyAdd(@a, @b);";
                 SqlCommand cmd = new SqlCommand(sql, conn);
 
                 cmd.Parameters.Add("@a", SqlDbType.Int);
@@ -22,12 +38,11 @@ namespace DB.Test
                 cmd.Parameters.Add("@b", SqlDbType.Int);
                 cmd.Parameters["@b"].Value = b;
 
-                conn.Open();
                 Int32 result = (Int32)cmd.ExecuteScalar();
 
                 Assert.Equal(expected, result);
-
             }
+
 
         }
 
@@ -37,8 +52,10 @@ namespace DB.Test
         [InlineData(-2, -5, -7)]
         public void usp_SimpleAdd_IntegerValues_Calculated(int a, int b, int expected)
         {
-            using (SqlConnection conn = new SqlConnection("Data Source=192.168.100.30,1433;Initial Catalog=SQLServerUT;Integrated Security=SSPI;"))
+
+            if (conn != null && conn.State == ConnectionState.Open)
             {
+
                 SqlCommand cmd = new SqlCommand("dbo.usp_SimpleAdd", conn);
                 cmd.CommandType = CommandType.StoredProcedure;
 
@@ -48,23 +65,22 @@ namespace DB.Test
                 cmd.Parameters.Add("@a", SqlDbType.Int).Value = a;
                 cmd.Parameters.Add("@b", SqlDbType.Int).Value = b;
 
-                conn.Open();
-                int rows = cmd.ExecuteNonQuery();               
+                int rows = cmd.ExecuteNonQuery();
                 int result = (int)cmd.Parameters["@Return Value"].Value;
 
                 Assert.Equal(expected, result);
-
+            
             }
+
         }
 
         [Fact]
         public void usp_ShowLogs_HappyPath_ShowLoginsOnly()
         {
 
-            using (SqlConnection conn = new SqlConnection("Data Source=192.168.100.30,1433;Initial Catalog=SQLServerUT;Integrated Security=SSPI;"))
-            {
 
-                conn.Open();
+            if (conn != null && conn.State == ConnectionState.Open)
+            {
 
                 SqlCommand cmd = new SqlCommand("TRUNCATE TABLE Users; TRUNCATE TABLE Events; TRUNCATE TABLE Logs", conn);
                 cmd.ExecuteNonQuery();
@@ -93,8 +109,8 @@ namespace DB.Test
                     }
 
                 }
-
             }
+
         }
